+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Builds after release 0.12.0
 
+#### Build 2105171
+
+-   Always copy MQTT payloads to prevent non-0-terminated strings
+-   Updated ArduinoJson to 6.18.0
+-   Added experimental support for `{"on":"t"}` to toggle on/off state via JSON
+
 #### Build 2105120
 
 -   Fixed possibility of non-0-terminated MQTT payloads